refactor: create core/ layer — types, constants, errors, registry
- Created src/core/constants.ts (HTTP codes, API types mapping)
- Created src/core/types.ts (Target, TransportResult, RawHeaders, etc.)
- Created src/core/errors.ts (SemaphoreError, ProviderSwitchNeeded)
- Created src/core/registry.ts (StateRegistry interface)
- Old files now re-export from core/ (backward compatible)
- All 580 tests pass
